+++ doc/conf.py
@@ -13,7 +13,7 @@ project = 'hyoga'
 copyright = '2021, Julien Seguinot'
 author = 'Julien Seguinot'
 version = '0.1 Akaishi'
-release = '0.1.0'
+release = '0.1.1'
 
 
 # -- General configuration ---------------------------------------------------